+++ assets/packages/core/tailwindcss-v3/play/intellisense.ts
@@ -157,7 +157,8 @@ async function classnameToCss(input) {
 
     let css = classes
         .map((className) => generate(className, context)).filter((x) => x !== null)
-        .map((value) => addPixelEquivalentsToValue(value, 16));
+        // .map((value) => addPixelEquivalentsToValue(value, 16))
+        ;
 
     return Array.isArray(css) ? css.join(" ") : css;
 }